pyproject.toml
[project]
name = "smart-detector-person"
version = "0.3.1" # Incremented version
description = "Person detection and recording system from RTSP/camera"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "ultralytics>=8.1.0",
    "opencv-python>=4.10.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
    "numpy>=1.20"
]

[project.scripts]
person-detector = "smart_detector.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/smart_detector/__init__.py"